requirements/base.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
django-tenants==3.7.0
psycopg2-binary==2.9.10

# Auth & security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-environ==0.11.2

# Async / background tasks
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Storage
django-storages==1.14.4
boto3==1.35.81

# Subscriptions / payments
stripe==11.3.0

# Utilities
Pillow==11.0.0
python-dateutil==2.9.0
